Add files via upload
This chess game is not refined, though I will likely choose to refine it in the future. The purpose of this program is to enable 2 users to actually play a game on a computer much like they would play with a chess board in real life. Nobody tells you if you won the chess game in real life, nobody will point out that the person you are playing with cheated except for you... and my program doesn't add such a feature.